:art: Split fixture types and values out
