+++ tests/fixtures/index.js
@@ -10,6 +10,8 @@ const slugify = require('underscore.string/slugify');
 const moment = require('moment');
 const _ = require('lodash');
 
+const futils = require('./utils');
+
 const utils = require('../../lib/utils');
 
 /**
@@ -61,132 +63,7 @@ const generate = (total, lang) => {
     eachOfType[types[i]] = 0;
     refs.content[slugify(types[i])] = [];
 
-    allTypes.push({
-      name: types[i],
-      id: slugify(types[i]),
-      description: ipsum({
-	count: Math.round(Math.random() * 6 + 1),
-	units: 'words',
-      }),
-      attributes: [
-	{
-	  name: 'Name',
-	  inputs: {
-	    text: {
-	      label: 'Name',
-	      type: 'text',
-	      id: uuid.v4(),
-	      name: 'name--text',
-	    },
-	  },
-	  id: `${slugify(types[i])}-name`,
-	  type: 'text',
-	},
-	{
-	  name: 'Textblock',
-	  inputs: {
-	    textblock: {
-	      label: 'Text',
-	      type: 'textarea',
-	      id: uuid.v4(),
-	      name: 'textblock--textarea',
-	    },
-	  },
-	  id: `${slugify(types[i])}-textblock`,
-	  type: 'textarea',
-	},
-	{
-	  name: 'Referencer',
-	  inputs: {
-	    referencer: {
-	      label: 'Text',
-	      type: 'reference',
-	      id: uuid.v4(),
-	      name: 'referencer--reference',
-	      settings: {
-		contentType: 'will-be-changed',
-	      },
-	      reference: true,
-	    },
-	  },
-	  id: `${slugify(types[i])}-referencer`,
-	  type: 'reference',
-	},
-	{
-	  name: 'Referencer Dual',
-	  inputs: {
-	    referencerdual1: {
-	      label: 'Referencer 1',
-	      type: 'reference',
-	      id: uuid.v4(),
-	      name: 'referencer1--reference',
-	      settings: {
-		contentType: 'will-be-changed',
-	      },
-	      reference: true,
-	    },
-	    referencerdual2: {
-	      label: 'Referencer 2',
-	      type: 'reference',
-	      id: uuid.v4(),
-	      name: 'referencer2--reference',
-	      settings: {
-		contentType: 'will-be-changed',
-	      },
-	      reference: true,
-	    },
-	  },
-	  id: `${slugify(types[i])}-referencer-dual`,
-	  type: 'reference',
-	},
-	{
-	  name: 'Referencer Repeating',
-	  inputs: {
-	    referencerrepeat: {
-	      label: 'Ref Repeat',
-	      type: 'reference',
-	      id: uuid.v4(),
-	      name: 'referencer-repeating--reference',
-	      settings: {
-		contentType: 'will-be-changed',
-	      },
-	      reference: true,
-	    },
-	  },
-	  id: `${slugify(types[i])}-referencer-repeating`,
-	  type: 'reference',
-	  repeatable: true,
-	},
-	{
-	  name: 'Referencer Dual Repeating',
-	  inputs: {
-	    referencerdualrepeat1: {
-	      label: 'Referencer 1',
-	      type: 'reference',
-	      id: uuid.v4(),
-	      name: 'referencer1--reference-repeating',
-	      settings: {
-		contentType: 'will-be-changed',
-	      },
-	      reference: true,
-	    },
-	    referencerdualrepeat2: {
-	      label: 'Referencer 2',
-	      type: 'reference',
-	      id: uuid.v4(),
-	      name: 'referencer2--reference-repeating',
-	      settings: {
-		contentType: 'will-be-changed',
-	      },
-	      reference: true,
-	    },
-	  },
-	  id: `${slugify(types[i])}-referencer-dual-repeating`,
-	  type: 'reference',
-	  repeatable: true,
-	},
-      ],
-    });
+    allTypes.push(futils.type(types[i]));
   }
 
   /**
@@ -253,53 +130,7 @@ const generate = (total, lang) => {
     // track number of pieces of content for each type
     eachOfType[type] += 1;
 
-    const values = {};
-    values[`${slugify(type)}-name`] = {
-      text: {
-	value: ipsum({
-	  count: 1,
-	  units: 'words',
-	  format: 'plain',
-	}),
-      },
-    };
-    values[`${slugify(type)}-textblock`] = {
-      textblock: {
-	value: ipsum({
-	  count: 2,
-	  units: 'paragraphs',
-	  format: 'plain',
-	  sentenceUpperBound: 5,
-	  paragraphUpperBound: 3,
-	}),
-      },
-    };
-    values[`${slugify(type)}-referencer`] = {
-      referencer: {
-	value: 'make-me-an-id',
-      },
-    };
-    values[`${slugify(type)}-referencer-dual`] = {
-      referencerdual1: {
-	value: 'make-me-an-id',
-      },
-      referencerdual2: {
-	value: 'make-me-an-id',
-      },
-    };
-    values[`${slugify(type)}-referencer-repeating`] = [{
-      referencerrepeat: {
-	value: 'make-me-an-id',
-      },
-    }];
-    values[`${slugify(type)}-referencer-dual-repeating`] = [{
-      referencerdualrepeat1: {
-	value: 'make-me-an-id',
-      },
-      referencerdualrepeat2: {
-	value: 'make-me-an-id',
-      },
-    }];
+    const values = futils.values(type);
 
     const item = {
       id,
